==> _setup/5b_generate_spp_map_list.Rmd <==
---
title: 'Generate IUCN spp map list'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/src/templates/ohara_hdr.html'
  pdf_document:
    toc: true
---

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(fig.width = 6, fig.height = 4, fig.path = 'Figs/',
                      echo = TRUE, message = FALSE, warning = FALSE)

library(raster)
library(rgeos)
source('https://raw.githubusercontent.com/oharac/src/master/R/common.R')  ###
  ### includes library(tidyverse); library(stringr); dir_M points to ohi directory
library(sf)

dir_git <- here()

source(file.path(dir_git, '_setup/common_fxns.R'))

dir_bli <- file.path(dir_M, 'git-annex/globalprep/_raw_data/birdlife_intl/d2019')
dir_shp <- file.path(dir_M, 'git-annex/globalprep/_raw_data/iucn_spp/d2019')
  ### in this folder are shapefiles at a taxonomic level, from
  ### IUCN spatial data download (version 2019-2).

```

# Summary

* Subpopulation polygons must be identified and rasterized separately from the parent polygon; this must be done by sciname and subpop fields since the polygon IDs are based upon the parent ID.
* Regional assessments need not be determined at this stage - the ID numbers match the global ID numbers (including subpops).

# Data sources

* IUCN species shapefiles:  IUCN. (2019). The IUCN Red List of Threatened Species. Version 2019-2. Retrieved August 2019, from http://www.iucnredlist.org
* BirdLife International shapefiles: BirdLife International and Handbook of the Birds of the World. (2018). Bird species distribution maps of the world. Version 7.0.  Available at http://datazone.birdlife.org/species/requestdis


# Methods

## Generate species map list with subpops

### Create subpopulation lookup to match maps with IUCN SIDs for subpops

From available shapefiles, reclassify subpopulation polygons into subpop ID numbers for matching to risk and trend. 

``` {r convert BLI to shp-dbf-etc}

### convert BirdLife Int'l .gdb to shapefile, after filtering to species in the "marine" list.
### Here we also standardize the column names to match the IUCN column names.

bli_marine_shp <- file.path(dir_bli, 'bli_marine_v2018.shp')

if(!file.exists(bli_marine_shp)) {
  
  bli_sf <- st_read(file.path(dir_bli, 'BOTW.gdb'))
  
  marine_spp <- read_csv(file.path(dir_data, sprintf('spp_marine_from_api_%s.csv', api_version)),
                         col_types = 'dcc')
  
  bli_marine_sf <- bli_sf %>%
    clean_df_names() %>%
    filter(sisid %in% marine_spp$iucn_sid)
  
  st_write(bli_marine_sf, bli_marine_shp)
  
  ### clean column names
  bli_marine_dbf <- str_replace(bli_marine_shp, '\\.shp$', '\\.dbf')
  
  att_table <- foreign::read.dbf(bli_marine_dbf) %>%
    setNames(names(bli_marine_sf)[1:(ncol(bli_marine_sf) - 1)]) %>%
    rename(iucn_sid = sisid, binomial = sciname, year = date, tax_comm = tax_com) %>%
    mutate(year = as.integer(as.character(year)))
    #  [1] "sisid"        "sciname"      "date"         "source"       "presence"     "origin"       "seasonal"    
    #  [8] "data_sens"    "sens_comm"    "compiler"     "tax_com"      "dist_com"     "reviewers"    "citation"    
    # [15] "version"      "shape_length" "shape_area"  
  ### from IUCN shapefiles:
    #  [1] "id_no"      "binomial"   "presence"   "origin"     "seasonal"   "compiler"   "year"       "citation"  
    #  [9] "source"     "dist_comm"  "island"     "subspecies" "subpop"     "legend"     "tax_comm"   "kingdom"   
    # [17] "phylum"     "class"      "order_"     "family"     "genus"      "code"       "marine"     "terrestial"
    # [25] "freshwater" "shape_Leng" "shape_Area"
  foreign::write.dbf(att_table, bli_marine_dbf)
}

```

Note that there appear to be no subpopulation polygons within the BirdLife International map set; there is no "subpop" column in that attribute table.  It appears that the individual files may not have "subpop" column either.

``` {r create_subpop_lookup}

iucn_dbfs <- list.files(dir_shp, pattern = '\\.dbf$', full.names = TRUE)
bli_dbfs  <- list.files(dir_bli, pattern = '\\.dbf$', full.names = TRUE)
shps_dbfs <- c(iucn_dbfs, bli_dbfs)

### This is all "marine" species by habitat and system:
marine_spp <- read_csv(file.path(dir_data, 
                                 sprintf('spp_marine_from_api_%s.csv', 
                                         api_version)),
                       col_types = 'dcc')

### This is *all* spp from the API, filtered down to those in marine_spp: 
api_spp <- read_csv(file.path(dir_o_anx, 'iucn', 
                              sprintf('spp_info_from_api_%s.csv', api_version)),
                    col_types = 'dccccccccc') %>%
  filter(iucn_sid %in% marine_spp$iucn_sid)

### Define helper function to cleanly get the info from DBF files,
### including colname checks
get_dbf <- function(x) { ### x <- shps_dbfs[1]
  spp_dbf_info <- foreign::read.dbf(x, as.is = TRUE) %>%
    clean_df_names()
  
  ### Individual files call the id number 'iucn_sid'; bli has been fixed to
  ### 'iucn_sid' as well; Red List Data Download files are 'id_no'... So:
  ### fix that here!
  names(spp_dbf_info)[names(spp_dbf_info) == 'id_no'] <- 'iucn_sid'

  ### if 'binomial', set to 'sciname'
  if(!'sciname' %in% names(spp_dbf_info)) {
    names(spp_dbf_info)[names(spp_dbf_info) == 'binomial'] <- 'sciname'
  }
  
  ### Some files (e.g. bli and individual files) don't have 'subpop':
  ### add it in as NAs.
  if(!'subpop' %in% names(spp_dbf_info)) spp_dbf_info$subpop <- NA
  
  ### some files don't have 'presence' field: if not present or NA,
  ### set to presence = 1.  Some have presence = 0?
  if(!'presence' %in% names(spp_dbf_info)) spp_dbf_info$presence <- 1
  
  spp_dbf_info <- spp_dbf_info %>%
    mutate(presence = ifelse(presence == 0, 1, presence))
  
  return(spp_dbf_info)
}

shps_df <- lapply(shps_dbfs, get_dbf) %>%
  setNames(shps_dbfs) %>%
  bind_rows(.id = 'dbf_file') %>%
  select(dbf_file, iucn_sid, sciname, presence, subpop) %>%
  filter(iucn_sid %in% marine_spp$iucn_sid)

shp_subpops <- shps_df %>%
  select(shp_iucn_sid = iucn_sid, sciname, shp_subpop = subpop) %>%
  filter(!is.na(shp_subpop)) %>%
  distinct() %>%
  mutate(shp_subpop_clean = str_replace(tolower(shp_subpop), ' subpopulation| ocean', ''),
         shp_subpop_clean = str_replace_all(shp_subpop_clean, '[^a-z]', ' '),
         shp_subpop_clean = str_trim(shp_subpop_clean),
         shp_subpop_clean = str_replace(shp_subpop_clean, 'noth', 'north'))
           ### fix typo

api_subpops <- api_spp %>%
  select(api_iucn_sid = iucn_sid, sciname, api_subpop = population) %>%
  filter(!is.na(api_subpop)) %>%
  filter(sciname %in% shp_subpops$sciname) %>%
  distinct() %>%
  mutate(api_subpop_clean = str_replace_all(tolower(api_subpop), ' subpopulation| ocean', ''),
         api_subpop_clean = str_replace_all(api_subpop_clean, '[^a-z]', ' ') %>% str_trim)

subpops_match_raw <- shp_subpops %>%
  full_join(api_subpops, by = 'sciname') %>%
  group_by(api_iucn_sid) %>%
  mutate(subpop_match = str_detect(shp_subpop_clean, api_subpop_clean),
         n_match = sum(subpop_match)) %>%
  filter(subpop_match | sum(subpop_match) == 0) %>%
  ungroup()

caretta_subpops <- subpops_match_raw %>%
  filter(shp_iucn_sid == 3897) %>%
  mutate(api_subpop_single = str_split(api_subpop_clean, ' ')) %>%
  unnest(api_subpop_single) %>%
  group_by(api_subpop_clean, shp_subpop_clean) %>%
  mutate(match = str_detect(shp_subpop_clean, api_subpop_single),
         n_match = sum(match),
         n_words = n()) %>%
  filter(sum(match) == n()) %>%
  ungroup() %>%
  select(shp_iucn_sid, sciname, shp_subpop, api_iucn_sid, api_subpop) %>%
  distinct()

subpops_match <- subpops_match_raw %>%
  filter(shp_iucn_sid != 3897 & subpop_match == TRUE) %>%
  select(shp_iucn_sid, sciname, shp_subpop, api_iucn_sid, api_subpop) %>%
  bind_rows(caretta_subpops)

### Check for missed matches:

# api_subpops$api_iucn_sid[!api_subpops$api_iucn_sid %in% subpops_match$api_iucn_sid]
### API subpop, not matched: 16369383 Tursiops truncatus 
###                                   Mediterranean subpopulation
### * OK - no polygon for this subpopulation.  Regional assessment?
### API subpop, not matched: 44187 Orcaella brevirostris 
###                                Malampaya Sound subpopulation
### * Doesn't look like there is a shape for this subpop... Sorry dude!

# shp_subpops[!shp_subpops$shp_subpop %in% subpops_match$shp_subpop, ]
### Shapefile subpop, not matched: 16285718 Chelonia mydas Hawaiian subpopulation
### * OK: the shapefile ID matches the API ID - will match automatically
### Balaena mysticetus in Bering-Chukchi-Beaufort Sea seems to not be 
###   a distinct assessment subpopulation, though there's a polygon.
### * Add to global population if not already?
### Dermochelys coriacea Southeast Atlantic is 46967848
### * Not in API subpops; Data deficient; ok to drop it
### Orcaella brevirostris Chilika Lake, Mekong River, Songkhla Lake
### * In shapefile, but not showing up in assessments
write_csv(subpops_match, 
          file.path(dir_setup, 
                    sprintf('int/subpops_match_api_to_shp_%s.csv', 
                            api_version)))

```

### get info on comprehensive assessments

Not all maps are for species that are included in comprehensively assessed taxonomic groups.  Get info on comprehensive assessment status for species from the API.

``` {r get comp assess status}
comp_file <- file.path(dir_data, 
                       sprintf('iucn_comp_assessed_%s.csv', api_version))
if(!file.exists(comp_file)) {
  comp_gps_url <- 'http://apiv3.iucnredlist.org/api/v3/comp-group/list?token=%s'
  comp_gps <- fromJSON(sprintf(comp_gps_url, api_key)) %>%
    .$result
  
  comp_spp_url <- 'http://apiv3.iucnredlist.org/api/v3/comp-group/getspecies/%s?token=%s'
  comp_spp <- mc_get_from_api(comp_spp_url, comp_gps$group_name, api_key) %>%
    select(iucn_sid = taxonid, sciname = scientific_name)
  
  write_csv(comp_spp, comp_file)
} else {
  comp_spp <- read_csv(comp_file,
                       col_types = 'dc')
}
```

### Join map list to subpop corrected list; remove duplicated spp IDs

From the species range map shapefiles, pull map info from the .dbf files.  For species listed in multiple files (e.g. sea snakes and reptiles), remove duplicates.  For species with polygons differentiated by subpopulation, adjust the iucn_sid to match the subpop info.

Add a flag for comprehensively-assessed spp - fix BONEFISH_TARPONS to all be assessed - either it's comprehensive or not - and add sea turtle species (6 of 7 assessed).

``` {r generate map list from dbfs}

iucn_dbfs  <- list.files(dir_shp, pattern = '\\.dbf$', full.names = TRUE)
bli_dbfs   <- list.files(dir_bli, pattern = '\\.dbf$', full.names = TRUE)

map_files  <- c(iucn_dbfs, bli_dbfs)


# spp_group_names <- map_files %>%
#   basename() %>%
#   str_replace_all('\\.dbf$|_PART_.+', '') %>%
#   unique()

map_info_raw <- lapply(map_files, get_dbf) %>%
  setNames(shps_dbfs) %>%
  bind_rows(.id = 'dbf_file') %>%
  select(shp_iucn_sid = iucn_sid, sciname, presence, subpop, dbf_file)

### Notes on presence, origin, seasonal fields:
### * presence = 5 is extinct; 4 = probably extinct; others are extant-ish or 
###   uncertain. We will drop field and include all polygons for now.
### * origin is native, introduced, etc.  We will drop this field and not
###   worry about origin.
### * seasonal is breeding/non-breeding/passage.  We will drop this field
###   and not worry about seasonality.

# subspp <- map_info %>%
#   filter(!is.na(subspecies)) %>%
#   select(id_no, code) %>%
#   distinct()
### No subspecies vary by IUCN code; drop field for simplicity
### and include all at species level

map_info_sans_dupes <- map_info_raw %>%
  group_by(shp_iucn_sid) %>%
  filter(!(str_detect(dbf_file, 'MARINEFISH|REPTILE|CORALS_PART_2_1|MANGROVES') & 
             length(unique(dbf_file)) > 1)) %>%
  filter(!(shp_iucn_sid == 196026 & str_detect(dbf_file, 'SEASNAKE'))) %>%
    ### in REPTILES, SEASNAKES, and HAGFISH - ditch any dupes
  ungroup() %>%
  distinct()
### NOTE: still has terrestrial mammals and reptiles in it...

subpops_match <- read_csv(file.path(dir_setup, 'int',
                                    sprintf('subpops_match_api_to_shp_%s.csv', 
                                            api_version)),
                          col_types = 'dccdc')

map_info_add_subpops <- map_info_sans_dupes %>%
  full_join(subpops_match, by = c('shp_iucn_sid', 'sciname', 'subpop' = 'shp_subpop')) %>%
  mutate(iucn_sid = ifelse(is.na(api_iucn_sid), shp_iucn_sid, api_iucn_sid)) %>%
  select(shp_iucn_sid, iucn_sid, sciname, presence, subpop, dbf_file) %>%
  distinct()

### filter to marine habitat species (to drop terrestrial reptiles e.g.)
marine_spp_ids <- read_csv(file.path(dir_data,
                                     sprintf('spp_marine_from_api_%s.csv', 
                                             api_version)),
                           col_types = 'dcc') %>%
  select(iucn_sid, max_depth)

marine_map_info <- map_info_add_subpops %>%
  inner_join(marine_spp_ids, by = 'iucn_sid') %>%
  mutate(comp_assessed = iucn_sid %in% comp_spp$iucn_sid)

# comp_assess_check <- marine_map_info %>%
#   mutate(dbf_short = basename(dbf_file)) %>%
#   group_by(dbf_short, comp_assessed) %>%
#   summarize(n_in_gp = n()) %>%
#   group_by(dbf_short) %>%
#   mutate(both = n() == 2)

### at this point, fix a few "comprehensively assessed" issues:
### * add turtles as comprehensive
### * some species from Spatial Data Download groups do not
###   show up as comprehensively assessed despite text on the IUCN site
turtles <- c('Dermochelys', 'Caretta', 'Lepidochelys', 
             'Natator', 'Chelonia', 'Eretmochelys') %>%
  paste0(collapse = '|')
assessed_gps <- c('BONEFISH', 'CORALS', 'SEACUCUMBERS', 
                  # 'CLUPEIFORM', 'SYNGATHIFORM', 
                  'bli_marine') %>%
  paste0(collapse = '|')
  ### NOTE: Damselfish removed at recommendation of Gina Ralph
  ### Others, assume that availability of maps on Spatial Data Download Page
  ### is indication of comprehensive assessment despite API call

marine_map_info <- marine_map_info %>%
  mutate(comp_assessed = ifelse(str_detect(dbf_file, assessed_gps), 
                                TRUE, comp_assessed),
         comp_assessed = ifelse(str_detect(sciname, turtles), 
                                TRUE, comp_assessed))

write_csv(marine_map_info, 
          file.path(dir_data,
                    sprintf('spp_marine_maps_%s.csv', api_version)))

```

``` {r generate table of species}

marine_map_info <- read_csv(file.path(dir_data,
    sprintf('spp_marine_maps_%s.csv', api_version))) %>%
  select(-presence) %>%
  distinct()

taxa_info <- read_csv(file.path(dir_o_anx, 'iucn', 
                                'spp_info_from_api_2018-1.csv')) %>%
  select(iucn_sid, kingdom:family)

cat_info <- read_csv(file.path(dir_data, 
                               'iucn_risk_current_2018-1.csv')) %>%
  select(iucn_sid, cat_score) %>%
  filter(!is.na(cat_score))

included_maps <- marine_map_info %>%
  filter(comp_assessed) %>%
  filter(iucn_sid %in% cat_info$iucn_sid) %>%
  select(iucn_sid, shp_iucn_sid, subpop, sciname, dbf_file) %>%
  left_join(taxa_info, by = 'iucn_sid') %>%
  mutate(spp_gp = basename(dbf_file),
         spp_gp = spp_gp %>% tolower() %>% str_replace('.dbf|_part.+', '')) %>%
  mutate(kingdom = tools::toTitleCase(tolower(kingdom)),
         phylum  = tools::toTitleCase(tolower(phylum)),
         order   = tools::toTitleCase(tolower(order)),
         class   = tools::toTitleCase(tolower(class)),
         family  = tools::toTitleCase(tolower(family)))

all_in_the_fam <- included_maps %>%
  group_by(kingdom, phylum, class, order, family, spp_gp) %>%
  summarize(species = length(unique(shp_iucn_sid)),
            subpops = sum(!is.na(subpop)))  

basic_sum <- included_maps %>%
  filter(!str_detect(tolower(class), 'actinopt|mammal|aves|rept')) %>%
  group_by(kingdom, phylum, class, spp_gp) %>%
  summarize(species = length(unique(shp_iucn_sid)),
            subpops = sum(!is.na(subpop)))

nonfish_sum <- included_maps %>%
  filter(str_detect(tolower(class), 'mammal|aves|rept')) %>%
  group_by(kingdom, phylum, class, order, spp_gp) %>%
  summarize(species = length(unique(shp_iucn_sid)),
            subpops = sum(!is.na(subpop)))

fish_sum <- included_maps %>%
  filter(str_detect(tolower(class), 'actinopt')) %>%
  group_by(kingdom, phylum, class, order, family, spp_gp) %>%
  summarize(species = length(unique(shp_iucn_sid)),
            subpops = sum(!is.na(subpop)))

all_sum <- bind_rows(fish_sum, nonfish_sum, basic_sum) %>%
  group_by(spp_gp, kingdom, phylum, class, order) %>%
  summarize(family = paste(unique(family), collapse = ', '),
            species = sum(species),
            subpops = sum(subpops)) %>%
  ungroup() %>%
  mutate(family = ifelse(family == 'NA', NA, family),
         subpops = ifelse(subpops == 0, NA, subpops))

write_csv(all_sum, file.path(dir_git, '_output', 
                             'spp_included_summary.csv'))
write_csv(all_in_the_fam, file.path(dir_git, '_output', 
                                    'spp_included_by_family.csv'))
```
